feat(explode): add explodedpayloades types
