+++ packages/explode/jest.config.ts
@@ -1,6 +1,7 @@
 import { JestConfigWithTsJest } from 'ts-jest'
 
 const config: JestConfigWithTsJest = {
+  prettierPath: null,
   testEnvironment: 'node',
   transform: {
     '^.+\\.tsx?$': ['ts-jest', { tsconfig: 'test/tsconfig.json' }],
